+++ internal/cluster/discovery.go
@@ -1,3 +1,17 @@
+// Copyright 2026 M. Javani
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 package cluster
 
 import (
@@ -16,7 +30,7 @@ import (
 // It returns immediately if the discovery kind is not HTTP.
 func (a *ClusterAgent) syncPeers(awg *sync.WaitGroup, tick time.Duration) {
 	defer awg.Done()
-	if a.discoveryKind != config.DiscoveryKindHttp { // Use your internal.DiscoveryKindHttp constant if available
+	if a.discoveryKind != config.DiscoveryKindHttp {
 		return
 	}